going remote

+++ Keepr.client/src/services/VaultKeepsService.js
@@ -13,7 +13,7 @@ class VaultKeepsService {
     logger.log('[createVaultKeep]', res.data)
     AppState.vaultKeeps.push(res.data)
   }
-  // NOTE I'm trying to accomplish gathering all the account's vault-keeps and replacing the keep in the Appstate with a keepViewModel so it has a vaultKeep Id on it to make for easy checking on the front end.  We'll see... YEAH IT WORKS
+  // NOTE I'm trying to accomplish gathering all the account's vault-keeps and replacing the keep in the Appstate with a keepViewModel so it has a vaultKeep Id on it to make for easy checking on the front end.  We'll see... it works but why is is needed? I don't think it is. Time to rework it.
   async getAllMyVaultsKeeps() {
     AppState.myVaults.forEach(async v => {
       const res = await api.get(`api/vaults/${v.id}/keeps`)
@@ -29,9 +29,13 @@ class VaultKeepsService {
   }
   async deleteVaultKeep(keepId) {
     const found = AppState.keeps.find(k => k.id = keepId)
-    const res = await api.delete(`api/vaultkeeps/${found.vaultKeepId}`)
-    logger.log('[deleteVaultKeep]', res.data)
-    AppState.keeps = AppState.keeps.filter(k => k.id != keepId)
+    debugger
+    logger.log(found.vaultKeepId)
+    if (found) {
+      const res = await api.delete(`api/vaultkeeps/${found.vaultKeepId}`)
+      logger.log('[deleteVaultKeep]', res.data)
+      AppState.keeps = AppState.keeps.filter(k => k.id != keepId)
+    }
   }
 }
 export const vaultKeepsService = new VaultKeepsService()
